feat: [TestRun] moved exception handlers into GlobalExceptionHandler
